- renamed Double.d to Float.d (per Don), renamed RFC1123.d to TimeStamp.d, and combined Atoi.d with Integer.d - converted Float.d, Integer.d, and TimeStamp.d to IFTI templates; changed all references found in the codebase - TimeStamp now supports wide chars - cleaned up some unittests - Atoi is now gone. Added some quick & dirty conversions to Integer.d

+++ tango/io/protocol/DisplayWriter.d
@@ -33,7 +33,7 @@ public  import  tango.io.protocol.model.IWriter;
         a convenient method of handling a variety of data types. 
         
         The DisplayWriter itself is a wrapper around the tango.text.convert 
-        package, which should be used directly as desired (Integer, Double, 
+        package, which should be used directly as desired (Integer, Float, 
         etc). The latter modules are home to a set of formatting-methods,
         making them convenient for ad-hoc application.
